Common - Move Navbar To Components
